Modularize extensions feature
