+++ desktop/pi-packages/index.cts
@@ -1,0 +1,3 @@
+export { searchPiPackages } from "./catalog.cts";
+export { listConfiguredPiPackages } from "./configured.cts";
+export { installPiPackage, removePiPackage } from "./mutations.cts";
