+++ src/app/features/extensions/types.ts
@@ -1,0 +1,13 @@
+export type ExtensionsViewProps = {
+  projectPath: string | null;
+  onSetProjectScopeActive: (active: boolean) => void;
+};
+
+export type InstallScope = "global" | "project";
+
+export type ManualSourceKind = "npm" | "git";
+
+export type PendingAction = {
+  kind: "install" | "remove";
+  source: string;
+};
